:recycle: color functions refactoring and Index.tsx is now managed inside index directoryh
