+++ app/pages/index/index_certificates/IndexCertificates.tsx
@@ -1,0 +1,12 @@
+import RowCard from "@/components/common/card/RowCard";
+import Timeline from "@/components/common/timeline/Timeline";
+
+export default function IndexCertificates() {
+  return (
+    <RowCard>
+      <Timeline>
+        자격증
+      </Timeline>
+    </RowCard>
+  );
+}
