+++ app/utils/color.ts
@@ -1,3 +1,5 @@
+const errorDefaultHex = "#999999";
+
 function getLuminance(hex: string) {
   const r = parseInt(hex.slice(1, 3), 16) / 255;
   const g = parseInt(hex.slice(3, 5), 16) / 255;
@@ -13,28 +15,25 @@ function getLuminance(hex: string) {
 
 /* 3자리 또는 6자리 HexCode를 포함한 Tailwind 클래스명에 대응 가능. */
 export function parseHexcode(text: string): string {
-    const match = text.match(/#[0-9A-Fa-f]+/);
-    if (!match) {
-      console.log(`invalid hex code : ${match}`);
-      return "";
-    }
-    
-    let hex = match[0];
+  const match = text.match(/#([a-fA-F0-9])+/);
+  if (!match) {
+    console.log(`invalid hex code : ${text}`);
+    return errorDefaultHex;
+  }
 
-    if(hex.length === 4)
-        hex = '#' + hex[1].repeat(2) + hex[2].repeat(2) + hex[3].repeat(2);
+  let hex = match[0];
 
-    return hex;
+  if (hex.length === 4)
+    hex = "#" + hex[1].repeat(2) + hex[2].repeat(2) + hex[3].repeat(2);
+  else if (hex.length !== 7) {
+    console.log(`invalid hex code : ${text}`);
+    hex = errorDefaultHex;
+  }
+
+  return hex;
 }
 
 export function isHexcodeLight(text: string) {
-  const match = text.match(/#[0-9A-Fa-f]{6}/);
-  if (!match) {
-    console.log(`invalid hex code : ${match}`);
-    return false;
-  }
-
-  const hex = parseHexcode(match[0]);
-
+  const hex = parseHexcode(text);
   return getLuminance(hex) > 0.5;
 }
